--- src/main/java/com/pnu/momeet/common/model/TokenPair.java
@@ -1,7 +1,0 @@
-package com.pnu.momeet.common.model;
-
-public record TokenPair(
-        String refreshToken,
-        String accessToken
-) {
-}
