+++ src/main/java/com/pnu/momeet/domain/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.pnu.momeet.domain.auth.dto;
+package com.pnu.momeet.domain.auth.dto.response;
 
 public record TokenResponse(
         String accessToken,
