+++ src/main/java/com/pnu/momeet/domain/member/dto/response/MemberInfo.java
@@ -1,4 +1,4 @@
-package com.pnu.momeet.domain.member.dto;
+package com.pnu.momeet.domain.member.dto.response;
 
 import com.pnu.momeet.domain.member.enums.Provider;
 
@@ -6,15 +6,16 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-public record MemberResponse(
+public record MemberInfo(
         UUID id,
         String email,
         Provider provider,
+        String password,
+        String providerId,
         List<String> roles,
         boolean enabled,
         boolean isAccountNonLocked,
-        LocalDateTime createdAt,
-        LocalDateTime updatedAt
+        LocalDateTime tokenIssuedAt
 ) {
 
 }
