+++ src/main/java/com/pnu/momeet/domain/member/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package com.pnu.momeet.domain.member.dto;
+package com.pnu.momeet.domain.member.dto.response;
 
 import com.pnu.momeet.domain.member.enums.Provider;
 
@@ -6,16 +6,15 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
-public record MemberInfo(
+public record MemberResponse(
         UUID id,
         String email,
         Provider provider,
-        String password,
-        String providerId,
         List<String> roles,
         boolean enabled,
         boolean isAccountNonLocked,
-        LocalDateTime tokenIssuedAt
+        LocalDateTime createdAt,
+        LocalDateTime updatedAt
 ) {
 
 }
