+++ src/main/java/com/pnu/momeet/domain/profile/controller/ProfileController.java
@@ -1,9 +1,9 @@
 package com.pnu.momeet.domain.profile.controller;
 
 import com.pnu.momeet.common.security.details.CustomUserDetails;
-import com.pnu.momeet.domain.profile.dto.ProfileCreateRequest;
-import com.pnu.momeet.domain.profile.dto.ProfileResponse;
-import com.pnu.momeet.domain.profile.dto.ProfileUpdateRequest;
+import com.pnu.momeet.domain.profile.dto.request.ProfileCreateRequest;
+import com.pnu.momeet.domain.profile.dto.response.ProfileResponse;
+import com.pnu.momeet.domain.profile.dto.request.ProfileUpdateRequest;
 import com.pnu.momeet.domain.profile.service.ProfileService;
 import jakarta.validation.Valid;
 import java.net.URI;
