+++ src/main/java/com/pnu/momeet/domain/profile/service/ProfileService.java
@@ -1,8 +1,8 @@
 package com.pnu.momeet.domain.profile.service;
 
-import com.pnu.momeet.domain.profile.dto.ProfileCreateRequest;
-import com.pnu.momeet.domain.profile.dto.ProfileResponse;
-import com.pnu.momeet.domain.profile.dto.ProfileUpdateRequest;
+import com.pnu.momeet.domain.profile.dto.request.ProfileCreateRequest;
+import com.pnu.momeet.domain.profile.dto.response.ProfileResponse;
+import com.pnu.momeet.domain.profile.dto.request.ProfileUpdateRequest;
 import com.pnu.momeet.domain.profile.entity.Profile;
 import com.pnu.momeet.domain.profile.enums.Gender;
 import com.pnu.momeet.domain.profile.mapper.EntityMapper;
