+++ src/test/java/com/pnu/momeet/e2e/sigungu/BaseSigunguTest.java
@@ -1,6 +1,6 @@
 package com.pnu.momeet.e2e.sigungu;
 
-import com.pnu.momeet.common.model.TokenPair;
+import com.pnu.momeet.domain.auth.dto.response.TokenResponse;
 import com.pnu.momeet.domain.auth.service.EmailAuthService;
 import com.pnu.momeet.domain.member.enums.Role;
 import com.pnu.momeet.e2e.BaseE2ETest;
@@ -14,7 +14,7 @@ import java.util.*;
 @Tag("sigungu")
 public abstract class BaseSigunguTest extends BaseE2ETest {
 
-    protected Map<Role, TokenPair> testTokens;
+    protected Map<Role, TokenResponse> testTokens;
 
     @Autowired
     protected EmailAuthService emailAuthService;
@@ -28,11 +28,11 @@ public abstract class BaseSigunguTest extends BaseE2ETest {
         testTokens.put(Role.ROLE_USER, emailAuthService.login(TEST_USER_EMAIL, TEST_USER_PASSWORD));
     }
 
-    protected TokenPair getToken() {
+    protected TokenResponse getToken() {
         return getToken(Role.ROLE_ADMIN);
     }
 
-    protected TokenPair getToken(Role role) {
+    protected TokenResponse getToken(Role role) {
         return testTokens.get(role);
     }
 }
